pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eidolon-vault"
version = "1.4.1"
description = "Persistent Scenario Intelligence Engine — multi‑agent simulation with continuous self‑learning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7,<9",
    "networkx>=3.2.1,<4",
    "pyyaml>=6.0.1,<7",
    "litellm>=1.44.2,<2.0",
    "trafilatura>=1.8.2",
    "pdfplumber>=0.11.0",
    "pypdf>=4.2.0",
    "requests>=2.32.3",
]

[project.optional-dependencies]
full = [
    "python-docx>=1.1.2",
]
dev = [
    "pytest>=8.0.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.5",
    "mypy>=1.9.0",
]

[project.scripts]
eidolon-vault = "eidolon_vault.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["eidolon_vault*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "S"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
